+++ src/middlewares/useInstrumentation.js
@@ -1,0 +1,9 @@
+import logger from '../logger';
+
+const useInstrumentation = (req, res, next) => {
+  req.instrumentation = logger;
+
+  next();
+};
+
+export default useInstrumentation;